pyproject.toml
[project]
name = "concierge-agent"
version = "0.1.0"
description = "Healthcare concierge agent using BeeAI and AgentStack"
requires-python = ">=3.13"
dependencies = [
    "agent-framework==1.0.0b251120",
    "agent-framework-a2a==1.0.0b251120",
    "agentstack-sdk==0.4.3",
    "beeai-framework[a2a]==0.1.74",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
concierge_agent = "concierge_agent.agentstack_concierge:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["concierge_agent"]
